+++ connect/connect-aws-dynamodb-sink/dynamodb-sink.sh
@@ -10,7 +10,7 @@ then
     exit 111
 fi
 
-if [ ! -z "$TAG_BASE" ] && version_gt $TAG_BASE "7.9.99" && [ ! -z "$CONNECTOR_TAG" ] && ! version_gt $CONNECTOR_TAG "1.4.99"
+if connect_cp_version_greater_than_8 && [ ! -z "$CONNECTOR_TAG" ] && ! version_gt $CONNECTOR_TAG "1.4.99"
 then
      logwarn "minimal supported connector version is 1.5.0 for CP 8.0"
      logwarn "see https://docs.confluent.io/platform/current/connect/supported-connector-version-8.0.html#supported-connector-versions-in-cp-8-0"
@@ -23,9 +23,41 @@ DYNAMODB_TABLE="pg${USER}dynamo${TAG}"
 DYNAMODB_ENDPOINT="https://dynamodb.$AWS_REGION.amazonaws.com"
 
 set +e
-log "Delete table, this might fail"
-aws dynamodb delete-table --table-name $DYNAMODB_TABLE --region $AWS_REGION
+aws dynamodb describe-table --table-name "$DYNAMODB_TABLE" --region $AWS_REGION --query 'Table.TableStatus' --output text 2>/dev/null
+if [ $? -eq 0 ]
+then
+  log "Delete table, this might fail"
+  aws dynamodb delete-table --table-name $DYNAMODB_TABLE --region $AWS_REGION
+  while true; do
+      aws dynamodb describe-table --table-name "$DYNAMODB_TABLE" --region $AWS_REGION --query 'Table.TableStatus' --output text 2>/dev/null
+      if [ $? -ne 0 ]
+      then
+          break
+      fi
+      sleep 5
+  done
+fi
 set -e
+
+log "Create dynamodb table $DYNAMODB_TABLE"
+aws dynamodb create-table \
+    --table-name "$DYNAMODB_TABLE" \
+    --attribute-definitions AttributeName=first_name,AttributeType=S AttributeName=last_name,AttributeType=S \
+    --key-schema AttributeName=first_name,KeyType=HASH AttributeName=last_name,KeyType=RANGE \
+    --provisioned-throughput ReadCapacityUnits=5,WriteCapacityUnits=5 \
+    --endpoint-url https://dynamodb.$AWS_REGION.amazonaws.com \
+    --tags Key=cflt_managed_by,Value=user Key=cflt_managed_id,Value="$USER"
+
+log "Waiting for table to be created"
+while true
+do
+    table_status=$(aws dynamodb describe-table --table-name "$DYNAMODB_TABLE" --region $AWS_REGION --query 'Table.TableStatus' --output text)
+    if [ "$table_status" == "ACTIVE" ]
+    then
+        break
+    fi
+    sleep 5
+done
 
 function cleanup_cloud_resources {
     set +e
@@ -39,16 +71,21 @@ PLAYGROUND_ENVIRONMENT=${PLAYGROUND_ENVIRONMENT:-"plaintext"}
 playground start-environment --environment "${PLAYGROUND_ENVIRONMENT}" --docker-compose-override-file "${PWD}/docker-compose.plaintext.yml"
 
 log "Sending messages to topic $DYNAMODB_TABLE"
-playground topic produce -t $DYNAMODB_TABLE --nb-messages 10 --forced-value '{"f1":"value%g"}' << 'EOF'
+playground topic produce -t $DYNAMODB_TABLE --nb-messages 10 << 'EOF'
 {
-  "type": "record",
-  "name": "myrecord",
   "fields": [
     {
-      "name": "f1",
+      "name": "first_name",
+      "type": "string"
+    },
+    {
+      "name": "last_name",
       "type": "string"
     }
-  ]
+  ],
+  "name": "Customer",
+  "namespace": "com.github.vdesabou",
+  "type": "record"
 }
 EOF
 
@@ -66,10 +103,11 @@ playground connector create-or-update --connector dynamodb-sink  << EOF
 }
 EOF
 
-log "Sleeping 120 seconds, waiting for table to be created"
-sleep 120
+sleep 10
+
+playground connector show-lag --max-wait 300
 
 log "Verify data is in DynamoDB"
 aws dynamodb scan --table-name $DYNAMODB_TABLE --region $AWS_REGION  > /tmp/result.log  2>&1
 cat /tmp/result.log
-grep "value1" /tmp/result.log
+grep "first_name" /tmp/result.log
